requirements.txt
dj-database-url==0.4.2
Django==1.11
djangorestframework==3.7.7
gunicorn==19.7.1
olefile==0.44
Pillow==4.1.1
pkg-resources==0.0.0
psycopg2==2.7.3
pytz==2018.3
